feat: add protection to privates pages

+++ controllers/admin.js
@@ -1,6 +1,9 @@
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
+	if (!req.session.isLoggedIn) {
+		return res.redirect('/login');
+	}
 	res.render('admin/edit-product', {
 		pageTitle: 'Add product',
 		path: '/admin/add-product',
@@ -19,7 +22,7 @@ exports.postAddProduct = (req, res, next) => {
 		imageUrl: imageUrl,
 		price: price,
 		description: descriptions,
-		userId: req.user
+		userId: req.user,
 	});
 	product
 		.save()
@@ -33,6 +36,9 @@ exports.postAddProduct = (req, res, next) => {
 };
 
 exports.getEditProduct = (req, res, next) => {
+	if (!req.session.isLoggedIn) {
+		return res.redirect('/login');
+	}
 	const editMode = req.query.edit;
 	if (!editMode) {
 		res.redirect('/');
@@ -94,6 +100,9 @@ exports.getProducts = (req, res, next) => {
 };
 
 exports.postDeleteProduct = (req, res, next) => {
+	if (!req.session.isLoggedIn) {
+		return res.redirect('/login');
+	}
 	const prodId = req.body.productId;
 	Product.findByIdAndRemove(prodId)
 		.then(() => {
